+++ runner/execute_test.go
@@ -1,34 +1,40 @@
-package fs
+package runner
 
-import "testing"
+import (
+	"os"
+	"testing"
+)
 
 func TestRun(t *testing.T) {
 
 	// Pass running a command without arguments
 	cmdNoArg := "ls"
-	_, errNoArg := Run(cmdNoArg)
+	cwd, _ := os.Getwd()
+	_, errNoArg := Run(cmdNoArg, cwd)
 	if errNoArg != nil {
 		t.Error(errNoArg)
 	}
 
 	// Pass by running a command with argument
 	cmdArg := "ls -l"
-	_, errArg := Run(cmdArg)
+	logF, errArg := Run(cmdArg, cwd)
 	if errArg != nil {
 		t.Error(errArg)
 	}
+	os.Remove(logF)
 
 	// Fail by passing an empty string
-	_, errEmpty := Run("")
+	_, errEmpty := Run("", cwd)
 	if errEmpty == nil {
 		t.Error("Expected error, got nil")
 	}
 
 	// Fail by passing a non-existing command
 	cmdNonExisting := "non-existing-command"
-	_, errNonExisting := Run(cmdNonExisting)
+	logF, errNonExisting := Run(cmdNonExisting, cwd)
 	if errNonExisting == nil {
 		t.Error("Expected error, got nil")
 	}
+	os.Remove(logF)
 
 }
